validateBGS and parseBGS working. All working
